+++ defaults/models/user.js
@@ -31,20 +31,10 @@ module.exports = {
 			allowNull: true,
 			secret: true
 		},
-		isSystem: {
-			type: Sequelize.BOOLEAN,
-			allowNull: false,
-			defaultValue: false
-		},
-		isRoot: {
-			type: Sequelize.BOOLEAN,
-			allowNull: false,
-			defaultValue: false
-		},
-		isPublic: {
-			type: Sequelize.BOOLEAN,
-			allowNull: false,
-			defaultValue: false
+		type: {
+			type: Sequelize.STRING(10),
+			allowNull: true,
+			unique: true
 		},
 		isActive: {
 			type: Sequelize.BOOLEAN,