+++ changelog.md
@@ -76,4 +76,5 @@
 ## Next
 
 * Change from `isRoot`, `isPublic` etc columns in user and role models to single column `type`
+* Create 'admin' and 'user' roles
 * Refactor db stocking code